+++ src/entries.test.ts
@@ -17,6 +17,7 @@ import * as v21 from "./v21";
 import * as v22 from "./v22";
 import * as v23 from "./v23";
 import * as v24 from "./v24";
+import * as v25 from "./v25";
 import * as v9 from "./v9";
 
 /**
@@ -44,6 +45,7 @@ const ENTRIES: Record<number, { BYPP_FORMAT_VERSION: number }> = {
   22: v22,
   23: v23,
   24: v24,
+  25: v25,
 };
 
 const SRC = dirname(fileURLToPath(import.meta.url));